requirements/base.txt
# Core data processing
pandas>=2.0,<3.0
pyarrow>=14.0
numpy>=1.24

# Geospatial
geopandas>=0.14
shapely>=2.0
fiona>=1.9
pyogrio>=0.7
rasterio>=1.3
rasterstats>=0.19

# Scale-up (census-parquet patterns)
dask>=2024.1
dask-geopandas>=0.3

# HTTP / downloads
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9

# Census API wrapper (npappin-wsu/census lineage)
census>=0.8.25

# Configuration
pyyaml>=6.0
python-dotenv>=1.0

# CLI
click>=8.1
